update full code in version2
