+++ requirements.txt
@@ -1,2 +1,2 @@
 --extra-index-url=https://artifacts.luxonis.com/artifactory/luxonis-python-snapshot-local
-depthai>=2.0
+depthai=2.2
